[Refactor] 레이어 분리
